multi_org_mgmt/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.3
django-import-export==4.1.1
django-widget-tweaks==1.5.0
python-decouple==3.8
sqlparse==0.5.2
tablib==3.5.0
